18/06

+++ finished/c_03/ex00/ft_strcmp.c
@@ -11,5 +11,5 @@ int	ft_strcmp(char *s1, char *s2)
 	i = 0;
 	while (s1[i] == s2[i] && s1[i] != '\0')
 		++i;
-	return (s1[i] - s2[i]);
+	return ((unsigned int)s1[i] - (unsigned int)s2[i]);
 }
